# PERF: tolist