The parent class implementation is faster.

```
dti = pd.date_range("2016-01-01", periods=10**5, freq="S")

%timeit dti.tolist()
286 ms ± 3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
62.8 ms ± 765 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```